Add files via upload
The Library Management System is designed to automate the
librarian’s work. The main objective of this project is to maintain
books of each category online. The various operations can be done
using this project such operations are inserting full details of book,
sorting stored books in ascending or descending orders, and also
delete unnecessary books. Search stored books with reference to
ISBN number.